+++ CHANGELOG.md
@@ -9,6 +9,8 @@
     - New homogeneous linearization syntax
     - Replaced bytecode compiler with (a little slower, but easier to maintain) pure AST interpreter
     - Fixed filenames in node's source: Now is the whole filename, with extension and dirs
+    - The source's file is now defined as a `fileName` property in Package so it doesn't need to be saved in each node
+    - New `isSynthetic` method for all nodes
 
 ## v2.3.0
 - Fixing building process in bad release